fix: quick appointment works with the focused day, CREATE APPOINTMENT button shows a tooltip and rename Header to HeaderView
