--- src/components/Calendar/Header/index.js
@@ -1,1 +1,0 @@
-export { HeaderView } from "./Header";
