+++ src/components/Calendar/index.js
@@ -1,4 +1,4 @@
-export { HeaderView } from "./Header";
+export { HeaderView } from "./HeaderView";
 export { DaysView } from "./Days";
 export { WeeksView } from "./Weeks";
 export { MonthsView } from "./Months";
