+++ src/pages/Room/RoomPage.jsx
@@ -2,6 +2,7 @@ import React from "react";
 import { CalendarContainer } from "./CalendarContainer";
 
 export const RoomPage = props => {
+  // TODO: roomId should be updated automatically, no need to reload window CalendarPageLogic
   const { roomId } = props.match.params;
   return <CalendarContainer URLRoomId={roomId} />;
 };
